Move ``BuildInfo`` to ``_build_info``
